resolve linter issues
