+++ src/runtime/transformation-dom-particle.ts
@@ -8,37 +8,42 @@
  * http://polymer.github.io/PATENTS.txt
  */
 'use strict';
-import { DomParticle } from './dom-particle.js';
+
+import {DomParticle} from './dom-particle.js';
+
 // Regex to separate style and template.
 const re = /<style>((?:.|[\r\n])*)<\/style>((?:.|[\r\n])*)/;
+
 /** @class TransformationDomParticle
  * Particle that does transformation stuff with DOM.
  */
 export class TransformationDomParticle extends DomParticle {
-    getTemplate(slotName) {
-        // TODO: add support for multiple slots.
-        return this._state.template;
-    }
-    getTemplateName(slotName) {
-        // TODO: add support for multiple slots.
-        return this._state.templateName;
-    }
-    render(props, state) {
-        return state.renderModel;
-    }
-    shouldRender(props, state) {
-        return Boolean((state.template || state.templateName) && state.renderModel);
-    }
-    renderHostedSlot(slotName, hostedSlotId, content) {
-        this.combineHostedTemplate(slotName, hostedSlotId, content);
-        this.combineHostedModel(slotName, hostedSlotId, content);
-    }
-    // abstract
-    combineHostedTemplate(slotName, hostedSlotId, content) { }
-    combineHostedModel(slotName, hostedSlotId, content) { }
-    // Helper methods that may be reused in transformation particles to combine hosted content.
-    static propsToItems(propsValues) {
-        return propsValues ? propsValues.map(({ rawData, id }) => Object.assign({}, rawData, { subId: id })) : [];
-    }
+  getTemplate(slotName) {
+    // TODO: add support for multiple slots.
+    return this._state.template;
+  }
+  getTemplateName(slotName) {
+    // TODO: add support for multiple slots.
+    return this._state.templateName;
+  }
+  render(props, state) {
+    return state.renderModel;
+  }
+  shouldRender(props, state) {
+    return Boolean((state.template || state.templateName) && state.renderModel);
+  }
+  renderHostedSlot(slotName, hostedSlotId, content) {
+    this.combineHostedTemplate(slotName, hostedSlotId, content);
+    this.combineHostedModel(slotName, hostedSlotId, content);
+  }
+  // abstract
+  combineHostedTemplate(slotName, hostedSlotId, content) { 
+  }
+  combineHostedModel(slotName, hostedSlotId, content) {
+  }
+  // Helper methods that may be reused in transformation particles to combine hosted content.
+  static propsToItems(propsValues) {
+    return propsValues ? propsValues.map(({rawData, id}) => ({...rawData, subId: id })) : [];
+  }
 }
 //# sourceMappingURL=transformation-dom-particle.js.map
